bws/requirements.txt
biopython==1.79
bs4==0.0.1
Django==4.0.4
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
drf-yasg==1.20.0
mysqlclient==2.0.3
python-dotenv==0.19.2